pyproject.toml
[project]
name = "copr-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "copr>=2.5",
    "pydantic-ai-slim[anthropic,mcp]>=1.73.0",
]
